Prune unused runtime
