+++ node/parachain/Cargo.toml
@@ -27,7 +27,7 @@ jsonrpc-core = "14.0.3"
 jsonrpc-pubsub = "14.0.3"
 
 # Parachain dependencies
-moonbase-runtime = { package = "parachain-runtime", path = "../../runtime/parachain", features = ["parachain"] }
+moonbeam-runtime = { path = "../../runtime/combined", features = ["parachain"] }
 
 # Substrate dependencies
 sp-runtime = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "rococo-branch" }
