+++ node/standalone/Cargo.toml
@@ -48,7 +48,7 @@ sp-runtime = { git = "https://github.com/paritytech/substrate.git", branch = "ro
 sc-basic-authorship = { git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
 sp-block-builder = { git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
 
-moonbeam-runtime = { package = "parachain-runtime", path = "../../runtime/parachain", features = ["standalone"] }
+moonbeam-runtime = {path = "../../runtime/combined", features = ["standalone"] }
 
 ethereum = { package = "pallet-ethereum", path = "../../vendor/frontier/frame/ethereum" }
 frontier-consensus = { path = "../../vendor/frontier/consensus" }
