--- runtime/standalone/Cargo.toml
@@ -1,84 +1,0 @@
-[package]
-authors = ['PureStake']
-edition = '2018'
-homepage = 'https://moonbeam.network'
-license = 'Unlicense'
-name = 'standalone-runtime'
-repository = 'https://github.com/PureStake/moonbeam/'
-version = '0.1.1'
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[dependencies]
-codec = { package = "parity-scale-codec", version = "1.3.0", default-features = false, features = ["derive"] }
-log = "0.4.8"
-hex-literal = "0.3.1"
-serde = { version = "1.0.101", optional = true, features = ["derive"] }
-
-pallet-aura = { default-features = false, package = "pallet-aura", git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-pallet-balances = { default-features = false, package = "pallet-balances", git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-frame-support = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-precompiles = { path = "../precompiles/", default-features = false }
-pallet-grandpa = { default-features = false, package = "pallet-grandpa", git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-pallet-randomness-collective-flip = { default-features = false, package = "pallet-randomness-collective-flip", git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-pallet-sudo = { default-features = false, package = "pallet-sudo", git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-frame-system = { default-features = false, package = "frame-system", git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-pallet-timestamp = { default-features = false, package = "pallet-timestamp", git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-pallet-transaction-payment = { default-features = false, package = "pallet-transaction-payment", git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-frame-evm = { default-features = false, package = "pallet-evm", git = "https://github.com/purestake/substrate", branch = "tgmichel-rococo-branch" }
-frame-executive = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-sp-api = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-sp-block-builder = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch"}
-sp-consensus-aura = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-sp-core = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-sp-inherents = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch"}
-sp-io = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-sp-offchain = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-sp-runtime = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-sp-session = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-sp-std = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-sp-transaction-pool = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-sp-version = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-frame-system-rpc-runtime-api = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-pallet-transaction-payment-rpc-runtime-api = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-
-frontier-rpc-primitives = { default-features = false, path = "../../vendor/frontier/rpc/primitives" }
-frame-ethereum = { version = "0.1.0", default-features = false, package = "pallet-ethereum", path = "../../vendor/frontier/frame/ethereum" }
-
-[build-dependencies]
-wasm-builder-runner = { version = "1.0.6", package = "substrate-wasm-builder-runner", git = "https://github.com/paritytech/substrate.git", branch = "rococo-branch" }
-
-[features]
-default = ["std"]
-std = [
-	"pallet-aura/std",
-	"pallet-balances/std",
-	"codec/std",
-	"frame-executive/std",
-	"frame-support/std",
-	"pallet-grandpa/std",
-	"pallet-randomness-collective-flip/std",
-	"serde",
-	"sp-api/std",
-	"sp-block-builder/std",
-	"sp-consensus-aura/std",
-	"sp-core/std",
-	"sp-inherents/std",
-	"sp-io/std",
-	"sp-offchain/std",
-	"sp-runtime/std",
-	"sp-session/std",
-	"sp-std/std",
-	"sp-transaction-pool/std",
-	"sp-version/std",
-	"pallet-sudo/std",
-	"frame-system/std",
-	"pallet-timestamp/std",
-	"pallet-transaction-payment/std",
-	"frame-ethereum/std",
-	"frame-evm/std",
-	"frame-system-rpc-runtime-api/std",
-	"pallet-transaction-payment-rpc-runtime-api/std",
-	"frontier-rpc-primitives/std",
-]
