--- runtime/standalone/src/constants.rs
@@ -1,83 +1,0 @@
-// Copyright 2019-2020 PureStake Inc.
-// This file is part of Moonbeam.
-
-// Moonbeam is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// Moonbeam is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with Moonbeam.  If not, see <http://www.gnu.org/licenses/>.
-
-//! A set of constant values used in substrate runtime.
-
-/// Money matters.
-pub mod currency {
-
-	pub type Balance = u128;
-
-	pub const MILLICENTS: Balance = 1_000_000_000;
-	pub const CENTS: Balance = 1_000 * MILLICENTS; // assume this is worth about a cent.
-	pub const DOLLARS: Balance = 100 * CENTS;
-
-	pub const MILLIGLMR: Balance = 1_000;
-	pub const CENTIGLMR: Balance = 1_000 * MILLIGLMR;
-	pub const GLMR: Balance = 100 * CENTIGLMR;
-}
-
-/// Time.
-pub mod time {
-
-	pub type BlockNumber = u32;
-	pub type Moment = u64;
-
-	/// Since BABE is probabilistic this is the average expected block time that
-	/// we are targetting. Blocks will be produced at a minimum duration defined
-	/// by `SLOT_DURATION`, but some slots will not be allocated to any
-	/// authority and hence no block will be produced. We expect to have this
-	/// block time on average following the defined slot duration and the value
-	/// of `c` configured for BABE (where `1 - c` represents the probability of
-	/// a slot being empty).
-	/// This value is only used indirectly to define the unit constants below
-	/// that are expressed in blocks. The rest of the code should use
-	/// `SLOT_DURATION` instead (like the Timestamp pallet for calculating the
-	/// minimum period).
-	///
-	/// If using BABE with secondary slots (default) then all of the slots will
-	/// always be assigned, in which case `MILLISECS_PER_BLOCK` and
-	/// `SLOT_DURATION` should have the same value.
-	///
-	/// <https://research.web3.foundation/en/latest/polkadot/BABE/Babe/#6-practical-results>
-	pub const MILLISECS_PER_BLOCK: Moment = 3000;
-	pub const SECS_PER_BLOCK: Moment = MILLISECS_PER_BLOCK / 1000;
-
-	pub const SLOT_DURATION: Moment = MILLISECS_PER_BLOCK;
-
-	// 1 in 4 blocks (on average, not counting collisions) will be primary BABE blocks.
-	pub const PRIMARY_PROBABILITY: (u64, u64) = (1, 4);
-
-	// Sessions are set low (1 per minute) for debugging purposes.
-	// As a block is produced every 3 seconds by configuration, a new session will occur every X blocks.
-	pub const EPOCH_DURATION_IN_BLOCKS: BlockNumber = 1 * MINUTES;
-	pub const EPOCH_DURATION_IN_SLOTS: u64 = {
-		const SLOT_FILL_RATE: f64 = MILLISECS_PER_BLOCK as f64 / SLOT_DURATION as f64;
-
-		(EPOCH_DURATION_IN_BLOCKS as f64 * SLOT_FILL_RATE) as u64
-	};
-
-	// These time units are defined in number of blocks.
-	pub const MINUTES: BlockNumber = 60 / (SECS_PER_BLOCK as BlockNumber);
-	pub const HOURS: BlockNumber = MINUTES * 60;
-	pub const DAYS: BlockNumber = HOURS * 24;
-
-	// epochs/session per era
-	pub const EPOCH_PER_ERA: u8 = 1;
-
-	// convenience year in milliseconds. TODO.
-	pub const MILLISECS_PER_YEAR: u64 = 3660 * 24 * 365 * 1000;
-}
